Make Signup component
