+++ src/App.js
@@ -21,6 +21,7 @@ import {
 import PrivateRoute from "./PrivateRoute";
 import Reset from "./pages/Reset";
 import Profile from "./pages/Profile";
+import EntryContextProvider from "./contexts/EntryContext";
 
 function App() {
   return (
@@ -35,9 +36,12 @@ function App() {
             <PrivateRoute path="/archive" component={Archive} />
             <PrivateRoute path="/search" component={Search} />
             <PrivateRoute path="/profile" component={Profile} />
-            <Route path="/login" component={Login} />
-            <Route path="/signup" component={Signup} />
-            <Route path="/reset" component={Reset} />
+            <EntryContextProvider>
+              <Route path="/login" component={Login} />
+              <Route path="/signup" component={Signup} />
+              <Route path="/reset" component={Reset} />
+            </EntryContextProvider>
+
             <Redirect to="/" component={Inbox} />
           </Switch>
         </TaskContextProvider>
