+++ src/contexts/AuthContext.js
@@ -59,7 +59,6 @@ const UserProvider = ({ children }) => {
         }
       } else {
         console.log("The user is not logged in");
-        history.push("/login");
       }
     });
   };
