--- src/screens/LoginScreen.js
@@ -1,8 +1,0 @@
-import React from "react";
-import Login from "../pages/Login";
-
-const LoginScreen = () => {
-  return <Login />;
-};
-
-export default LoginScreen;
