--- src/screens/SignupScreen.js
@@ -1,8 +1,0 @@
-import React from "react";
-import { Signup } from "../pages";
-
-const LoginScreen = () => {
-  return <Signup />;
-};
-
-export default LoginScreen;
